+++ share/eds/blockstore.go
@@ -4,16 +4,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io"
 
 	bstore "github.com/ipfs/boxo/blockstore"
 	dshelp "github.com/ipfs/boxo/datastore/dshelp"
 	blocks "github.com/ipfs/go-block-format"
 	"github.com/ipfs/go-cid"
 	"github.com/ipfs/go-datastore"
+	"github.com/ipfs/go-datastore/namespace"
 	ipld "github.com/ipfs/go-ipld-format"
-
-	"github.com/celestiaorg/celestia-node/share/eds/cache"
 )
 
 var _ bstore.Blockstore = (*blockstore)(nil)
@@ -33,14 +31,12 @@ var (
 // implementation to allow for the blockstore operations to be routed to the underlying stores.
 type blockstore struct {
 	store *Store
-	cache cache.Cache
 	ds    datastore.Batching
 }
 
-func newBlockstore(store *Store, cache cache.Cache, ds datastore.Batching) *blockstore {
+func newBlockstore(store *Store, ds datastore.Batching) *blockstore {
 	return &blockstore{
 		store: store,
-		cache: cache,
 		ds:    namespace.Wrap(ds, blockstoreCacheKey),
 	}
 }
@@ -156,28 +152,17 @@ func (bs *blockstore) getReadOnlyBlockstore(ctx context.Context, cid cid.Cid) (*
 		return nil, fmt.Errorf("failed to find shards containing multihash: %w", err)
 	}
 
-	// check if cache contains any of accessors
+	// check if any if multicache contains an accessor
 	shardKey := keys[0]
-	if accessor, err := bs.store.cache.Get(shardKey); err == nil {
+	accessor, err := bs.store.cache.Get(shardKey)
+	if err == nil {
 		return blockstoreCloser(accessor)
 	}
 
 	// load accessor to the blockstore cache and use it as blockstoreCloser
-	accessor, err := bs.cache.GetOrLoad(ctx, shardKey, bs.store.getAccessor)
+	accessor, err = bs.store.cache.Second().GetOrLoad(ctx, shardKey, bs.store.getAccessor)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get accessor for shard %s: %w", shardKey, err)
 	}
 	return blockstoreCloser(accessor)
-}
-
-// blockstoreCloser constructs new BlockstoreCloser from cache.Accessor
-func blockstoreCloser(ac cache.Accessor) (*BlockstoreCloser, error) {
-	bs, err := ac.Blockstore()
-	if err != nil {
-		return nil, fmt.Errorf("eds/store: failed to get blockstore: %w", err)
-	}
-	return &BlockstoreCloser{
-		ReadBlockstore: bs,
-		Closer:         ac,
-	}, nil
 }
