+++ share/eds/cache/multicache.go
@@ -1,13 +1,10 @@
 package cache
 
 import (
-	"context"
 	"errors"
 
 	"github.com/filecoin-project/dagstore/shard"
 )
-
-var _ Cache = (*MultiCache)(nil)
 
 // MultiCache represents a Cache that looks into multiple caches one by one.
 type MultiCache struct {
@@ -31,25 +28,19 @@ func (mc *MultiCache) Get(key shard.Key) (Accessor, error) {
 	return mc.second.Get(key)
 }
 
-// GetOrLoad attempts to get an item from all caches, and if not found, invokes
-// the provided loader function to load it into one of the caches.
-func (mc *MultiCache) GetOrLoad(
-	ctx context.Context,
-	key shard.Key,
-	loader func(context.Context, shard.Key) (Accessor, error),
-) (Accessor, error) {
-	ac, err := mc.first.GetOrLoad(ctx, key, loader)
-	if err == nil {
-		return ac, nil
-	}
-	return mc.second.GetOrLoad(ctx, key, loader)
-}
-
 // Remove removes an item from all underlying caches
 func (mc *MultiCache) Remove(key shard.Key) error {
 	err1 := mc.first.Remove(key)
 	err2 := mc.second.Remove(key)
 	return errors.Join(err1, err2)
+}
+
+func (mc *MultiCache) First() Cache {
+	return mc.first
+}
+
+func (mc *MultiCache) Second() Cache {
+	return mc.second
 }
 
 func (mc *MultiCache) EnableMetrics() error {
